+++ src/App.jsx
@@ -9,6 +9,7 @@ import ControlsReview from './views/ControlsReview'
 import ModeSelectorCompare from './views/ModeSelectorCompare'
 import EnergySelectReview from './views/EnergySelectReview'
 import SoftkeyReview from './views/SoftkeyReview'
+import TypographyReview from './views/TypographyReview'
 
 export default function App() {
   const [params] = useSearchParams()
@@ -28,6 +29,7 @@ export default function App() {
         <Route path="/mode-selector-compare" element={<ModeSelectorCompare />} />
         <Route path="/energy-select-review" element={<EnergySelectReview />} />
         <Route path="/softkey-review" element={<SoftkeyReview />} />
+        <Route path="/typography-review" element={<TypographyReview />} />
         <Route path="*" element={<Navigate to="/" replace />} />
       </Routes>
     </SimulatorProvider>